Validación de roles y permisos

+++ app/Http/Controllers/ClienteController.php
@@ -8,11 +8,7 @@ use Illuminate\Support\Facades\Hash;
 
 class ClienteController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+  
     public function __construct()
     {
         $this->middleware('auth');
@@ -21,6 +17,11 @@ class ClienteController extends Controller
     public function index()
     {
         //
+
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
+        
         $datos['usuarios'] = Usuario::where('role', '=', '3')
                                     ->where('estado', '=', '1')
                                     ->orderBy('role', 'asc')
@@ -32,28 +33,25 @@ class ClienteController extends Controller
         return view('cliente.index', $datos, $contador);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         //
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
+
         $contador = (new UsuarioController)->count();
 
         return view('cliente.create', $contador);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         //
+
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
 
         $datosUsuario = $this->validate(request(), [
             'email' => 'email|required|string|unique:usuario',
@@ -74,25 +72,13 @@ class ClienteController extends Controller
         return redirect('cliente')->with('Mensaje', 'Cliente agregado correctamente.' );
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Usuario  $usuario
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Usuario $usuario)
-    {
-        //
-    }
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Usuario  $usuario
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         //
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
+
         $usuarios = Usuario::findOrFail($id);
 
         $contador = (new UsuarioController)->count();
@@ -100,16 +86,12 @@ class ClienteController extends Controller
         return view('cliente.edit', compact('usuarios'), $contador);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Usuario $usuario
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
         //
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
 
         $datosUsuario = request()->except(['_token', '_method']);
 
@@ -130,15 +112,13 @@ class ClienteController extends Controller
         return redirect('cliente')->with('Mensaje', 'Usuario actualizado correctamente.' );
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Usuario $usuario
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
         //
+        if (auth()->user()->role == 3) {
+            return redirect('informe');
+        }
+        
         Usuario::destroy($id);Usuario::where('id', '=', $id)->update([
             'estado' => 0
         ]);
